+++ packages/components/organisms/f-header/src/components/_tests/Header.test.js
@@ -6,7 +6,7 @@ const $style = {
     'c-header-container': 'c-header-container',
     'c-header--transparent': 'c-header--transparent',
     'c-header--gradient': 'c-header--gradient',
-    'c-header--search': 'c-header--search'
+    'c-header--disappearingWhite': 'c-header--disappearingWhite'
 };
 
 describe('Header', () => {